// frontend/src/api/generated/models/AddExpenseTemplateRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Rental API
 * API for rental management
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ExpenseCategory } from './ExpenseCategory';
import {
    ExpenseCategoryFromJSON,
    ExpenseCategoryFromJSONTyped,
    ExpenseCategoryToJSON,
} from './ExpenseCategory';

/**
 * 
 * @export
 * @interface AddExpenseTemplateRequest
 */
export interface AddExpenseTemplateRequest {
    /**
     * 
     * @type {number}
     * @memberof AddExpenseTemplateRequest
     */
    apartmentId?: number | null;
    /**
     * 
     * @type {number}
     * @memberof AddExpenseTemplateRequest
     */
    roomId?: number | null;
    /**
     * 
     * @type {number}
     * @memberof AddExpenseTemplateRequest
     */
    amount: number;
    /**
     * 
     * @type {ExpenseCategory}
     * @memberof AddExpenseTemplateRequest
     */
    category: ExpenseCategory;
    /**
     * 
     * @type {string}
     * @memberof AddExpenseTemplateRequest
     */
    expenseDate: string;
}



/**
 * Check if a given object implements the AddExpenseTemplateRequest interface.
 */
export function instanceOfAddExpenseTemplateRequest(value: object): value is AddExpenseTemplateRequest {
    if (!('amount' in value) || value['amount'] === undefined) return false;
    if (!('category' in value) || value['category'] === undefined) return false;
    if (!('expenseDate' in value) || value['expenseDate'] === undefined) return false;
    return true;
}

export function AddExpenseTemplateRequestFromJSON(json: any): AddExpenseTemplateRequest {
    return AddExpenseTemplateRequestFromJSONTyped(json, false);
}

export function AddExpenseTemplateRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): AddExpenseTemplateRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'apartmentId': json['apartmentId'] == null ? undefined : json['apartmentId'],
        'roomId': json['roomId'] == null ? undefined : json['roomId'],
        'amount': json['amount'],
        'category': ExpenseCategoryFromJSON(json['category']),
        'expenseDate': json['expenseDate'],
    };
}

export function AddExpenseTemplateRequestToJSON(value?: AddExpenseTemplateRequest | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'apartmentId': value['apartmentId'],
        'roomId': value['roomId'],
        'amount': value['amount'],
        'category': ExpenseCategoryToJSON(value['category']),
        'expenseDate': value['expenseDate'],
    };
}