// frontend/src/api/generated/models/ExpenseCategory.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Rental API
 * API for rental management
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ExpenseCategory = {
    OwnerRent: 'OWNER_RENT',
    UtilityWaterCold: 'UTILITY_WATER_COLD',
    UtilityWaterHot: 'UTILITY_WATER_HOT',
    UtilityElectricity: 'UTILITY_ELECTRICITY',
    UtilityGas: 'UTILITY_GAS',
    TaxZus: 'TAX_ZUS',
    TaxPit: 'TAX_PIT',
    Other: 'OTHER'
} as const;
export type ExpenseCategory = typeof ExpenseCategory[keyof typeof ExpenseCategory];


export function instanceOfExpenseCategory(value: any): boolean {
    for (const key in ExpenseCategory) {
        if (Object.prototype.hasOwnProperty.call(ExpenseCategory, key)) {
            if (ExpenseCategory[key as keyof typeof ExpenseCategory] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ExpenseCategoryFromJSON(json: any): ExpenseCategory {
    return ExpenseCategoryFromJSONTyped(json, false);
}

export function ExpenseCategoryFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExpenseCategory {
    return json as ExpenseCategory;
}

export function ExpenseCategoryToJSON(value?: ExpenseCategory | null): any {
    return value as any;
}