// frontend/src/api/generated/models/UpdateExpenseTemplate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Rental API
 * API for rental management
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ExpenseCategory } from './ExpenseCategory';
import {
    ExpenseCategoryFromJSON,
    ExpenseCategoryFromJSONTyped,
    ExpenseCategoryToJSON,
} from './ExpenseCategory';

/**
 * 
 * @export
 * @interface UpdateExpenseTemplate
 */
export interface UpdateExpenseTemplate {
    /**
     * 
     * @type {number}
     * @memberof UpdateExpenseTemplate
     */
    apartmentId?: number | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateExpenseTemplate
     */
    roomId?: number | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateExpenseTemplate
     */
    amount?: number | null;
    /**
     * 
     * @type {ExpenseCategory}
     * @memberof UpdateExpenseTemplate
     */
    category?: ExpenseCategory | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateExpenseTemplate
     */
    expenseDate?: string | null;
}



/**
 * Check if a given object implements the UpdateExpenseTemplate interface.
 */
export function instanceOfUpdateExpenseTemplate(value: object): value is UpdateExpenseTemplate {
    return true;
}

export function UpdateExpenseTemplateFromJSON(json: any): UpdateExpenseTemplate {
    return UpdateExpenseTemplateFromJSONTyped(json, false);
}

export function UpdateExpenseTemplateFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateExpenseTemplate {
    if (json == null) {
        return json;
    }
    return {
        
        'apartmentId': json['apartmentId'] == null ? undefined : json['apartmentId'],
        'roomId': json['roomId'] == null ? undefined : json['roomId'],
        'amount': json['amount'] == null ? undefined : json['amount'],
        'category': json['category'] == null ? undefined : ExpenseCategoryFromJSON(json['category']),
        'expenseDate': json['expenseDate'] == null ? undefined : json['expenseDate'],
    };
}

export function UpdateExpenseTemplateToJSON(value?: UpdateExpenseTemplate | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'apartmentId': value['apartmentId'],
        'roomId': value['roomId'],
        'amount': value['amount'],
        'category': ExpenseCategoryToJSON(value['category']),
        'expenseDate': value['expenseDate'],
    };
}